fix: ExtensionFS (closes #161)

+++ app/Console/Commands/BlueprintFramework/Test/StorageCommand.php
@@ -8,7 +8,7 @@ use Illuminate\Support\Facades\Storage;
 class StorageCommand extends Command
 {
   protected $description = 'Perform a test operation on a Laravel storage filesystem';
-  protected $signature = 'bp:test:storage {--disk=}';
+  protected $signature = 'bp:test:storage {--disk=} {--clean=}';
 
   public function __construct()
   {
@@ -20,8 +20,12 @@ class StorageCommand extends Command
     $this->info('This command will perform a test operation on a filesystem disk');
 
     $disk = $this->option('disk') ?? $this->ask('Which disk to test?');
+    $clean = $this->option('clean') ?? $this->choice('Clean up file afterwards?');
 
     Storage::disk($disk)->put('bp_fs_test.txt', 'hi :)');
-    Storage::disk($disk)->delete('bp_fs_test.txt');
+
+    if ($clean == 'Yes') {
+      Storage::disk($disk)->delete('bp_fs_test.txt');
+    }
   }
 }
